refresh from Dev

+++ backend/src/service/transfer-service.ts
@@ -9,12 +9,15 @@ import extractsTransferInfo from "../utils/extractsTransferInfo";
 import createFolder from "../utils/createFolder";
 import extractsDigitalFileList from "../utils/extractsDigitalFileList";
 import multer from "multer";
+import S3ClientService from "./s3Client-service";
 
 export default class TransferService {
   private transferRepository: TransferRepository;
+  private s3ClientService: S3ClientService;
 
   constructor() {
     this.transferRepository = new TransferRepository();
+    this.s3ClientService = new S3ClientService();
   }
 
   async getTransfers(): Promise<ITransfer[] | null> {
@@ -49,37 +52,36 @@ export default class TransferService {
   }
 
   async createTransferMetaData(
-       filePath: string
+    filePath: string
   ) {
-        const transferData = await extractsFromAra66x(filePath);
-        const accessionNumber = transferData?.accession || " ";
-        const applicationNumber = transferData?.application || " ";
-        const producerMinistry = transferData?.ministry || " ";
-        const producerBranch = transferData?.branch || " ";
+    const transferData = await extractsFromAra66x(filePath);
+    const accessionNumber = transferData?.accession || " ";
+    const applicationNumber = transferData?.application || " ";
+    const producerMinistry = transferData?.ministry || " ";
+    const producerBranch = transferData?.branch || " ";
 
-        const transferMetaData: ITransfer = {
-          "accessionNumber": accessionNumber,
-          "applicationNumber": applicationNumber,
-          "producerMinistry": producerMinistry,
-          "producerBranch": producerBranch
-        }
-        //check if the transfer exists before creating one
-        var transfer = await this.getTransferByKeysNumbers(accessionNumber, applicationNumber);
-        if(transfer && transfer._id)
-          {
-            transfer.producerMinistry = producerMinistry;
-            transfer.producerBranch = producerBranch;
-            this.updateTransfer(transfer._id?.toString() , transfer);
-          }
-        const newTransfer =await this.createTransfer(transferMetaData);
-        return newTransfer;
+    const transferMetaData: ITransfer = {
+      "accessionNumber": accessionNumber,
+      "applicationNumber": applicationNumber,
+      "producerMinistry": producerMinistry,
+      "producerBranch": producerBranch
+    }
+    //check if the transfer exists before creating one
+    var transfer = await this.getTransferByKeysNumbers(accessionNumber, applicationNumber);
+    if (transfer && transfer._id) {
+      transfer.producerMinistry = producerMinistry;
+      transfer.producerBranch = producerBranch;
+      this.updateTransfer(transfer._id?.toString(), transfer);
+    }
+    const newTransfer = await this.createTransfer(transferMetaData);
+    return newTransfer;
   }
 
   async extractsTransferInfo(
     filePath: string
   ) {
-      const transferData = await extractsTransferInfo(filePath);
-      return transferData;
+    const transferData = await extractsTransferInfo(filePath);
+    return transferData;
   }
 
   async createTransfer(transfer: ITransfer): Promise<ITransfer | null> {
@@ -100,7 +102,32 @@ export default class TransferService {
     }
   }
 
+  // async deleteTransfer(transferId: any) {
+  //   return await this.transferRepository.deleteTransfer(transferId);
+  // }
+
   async deleteTransfer(transferId: any) {
+    // Find the transfer
+    const transfer = await this.transferRepository.getTransfersById(transferId);
+    if (!transfer) {
+      throw new Error('Transfer not found');
+    }
+
+    // Construct the folder path
+    const folderPath = 'Transfers/' + transfer.accessionNumber + '-' + transfer.applicationNumber + '/';
+    console.log("deleteTransfer:" + folderPath)
+    // Delete the folder in S3
+    await this.s3ClientService.deleteFolder(folderPath);
+
+    // Delete the transfer
     return await this.transferRepository.deleteTransfer(transferId);
+  }
+
+  async createPSPs(transferId: any) {
+    const transfer = await this.transferRepository.getTransfersById(transferId);
+    if (!transfer) {
+      throw new Error('Transfer not found');
+    }
+
   }
 }
